requirements.txt
# deep learning framework
torch>=1.10.0
torchvision>=0.11.0

# array and image processing
numpy>=1.19.5
Pillow>=8.2.0
opencv-python>=4.5.3.56

# progress bar
tqdm>=4.62.0

# semantic segmentation
segmentation-models-pytorch>=0.2.1
timm>=0.6.5

# Conditional Random Fields
pydensecrf>=1.0.0
